+++ jvm/src/main/scala/eval/directive/OpDirectives.scala
@@ -18,10 +18,9 @@ import scala.reflect.ClassTag
 import scala.concurrent.duration._
 import scala.util.Try
 
-
 object OpDirectives {
 
-  private def asInstanceOfOption[T: ClassTag](o: Any): Option[T] = 
+  private def asInstanceOfOption[T: ClassTag](o: Any): Option[T] =
     Some(o) collect { case m: T => m}
 
   private def doubleResults(grouped: Map[MamlKind, Seq[Result]]): Interpreted[List[Double]] =
@@ -182,30 +181,56 @@ object OpDirectives {
       .andThen({ results => Valid(DoubleResult(results.reduce(_ max _))) })
   }
 
-  val maxTile = Directive { case (a@Max(_), childResults) if (a.kind == MamlKind.Image) =>
-  val grouped = childResults.groupBy(_.kind)
+  val maxTile = Directive { case (a@Max(_), childResults) if a.kind == MamlKind.Image =>
+    val grouped = childResults.groupBy(_.kind)
 
-  val scalarMax: Interpreted[Option[Double]] =
-    (doubleResults(grouped), intResults(grouped)).mapN { case (dbls, ints) =>
-      (Try(dbls.max).toOption, Try(ints.max).toOption) match {
-        case (Some(dbl), Some(int)) => Some(dbl max int)
-        case (None, Some(int)) => Some(int)
-        case (Some(dbl), None) => Some(dbl)
-        case _ => None
+    val scalarMax: Interpreted[Option[Double]] =
+      (doubleResults(grouped), intResults(grouped)).mapN { case (dbls, ints) =>
+        (Try(dbls.max).toOption, Try(ints.max).toOption) match {
+          case (Some(dbl), Some(int)) => Some(dbl max int)
+          case (None, Some(int)) => Some(int)
+          case (Some(dbl), None) => Some(dbl)
+          case _ => None
+        }
+      }
+
+    (imageResults(grouped), scalarMax).mapN({ case (tiles, maximum) =>
+      val tileMax = tiles.reduce({ (lt1: LazyMultibandRaster, lt2: LazyMultibandRaster) =>
+        lt1.dualCombine(lt2, {_ max _}, {_ max _})
+      })
+      maximum match {
+        case Some(scalarMax) =>
+          ImageResult(tileMax.dualMap({ i: Int => i max scalarMax.toInt }, { i: Double => i max scalarMax }))
+        case None =>
+          ImageResult(tileMax)
+      }
+    })
+  }
+
+  val rgbTile = Directive { case (a @ RGB(_, rb, gb, bb), childResults) if a.kind == MamlKind.Image =>
+    val grouped = childResults.groupBy(_.kind)
+
+    imageResults(grouped).map { tiles => tiles.take(3) match {
+        case r :: g :: b :: Nil =>
+          ImageResult(LazyMultibandRaster(Map("0" -> r.bands(rb), "1" -> g.bands(gb), "2" -> b.bands(bb))))
+        case list => ImageResult(list.reduce { (lt1: LazyMultibandRaster, lt2: LazyMultibandRaster) =>
+          LazyMultibandRaster(lt1.bands ++ lt2.bands)
+        })
       }
     }
+  }
 
-  (imageResults(grouped), scalarMax).mapN({ case (tiles, maximum) =>
-    val tileMax = tiles.reduce({ (lt1: LazyMultibandRaster, lt2: LazyMultibandRaster) =>
-      lt1.dualCombine(lt2, {_ max _}, {_ max _})
-    })
-    maximum match {
-      case Some(scalarMax) =>
-        ImageResult(tileMax.dualMap({ i: Int => i max scalarMax.toInt }, { i: Double => i max scalarMax }))
-      case None =>
-        ImageResult(tileMax)
+  val assembleTile = Directive { case (a @ Assemble(_), childResults) if a.kind == MamlKind.Image =>
+    val grouped = childResults.groupBy(_.kind)
+
+    imageResults(grouped).map { tiles =>
+      val bands =
+        tiles.zipWithIndex.map { case (tile, index) =>
+          index.toString -> tile.bands("0")
+        }
+
+      ImageResult(LazyMultibandRaster(bands.toMap))
     }
-  })
   }
 
   val minDouble = Directive { case (a@Min(_), childResults) if (a.kind == MamlKind.Double) =>
@@ -391,4 +416,3 @@ object OpDirectives {
     Valid(results)
   }
 }
-